// src/app/pages/modulos/components/modulos-table/modulos.interface.ts
export interface ModulosInterface {
    claveauth?: string;
    nicknameauth?: string;
    usuarioauth?: string;
    idmodulo?: number;
    descripcion: string;
    direccion: string;
    medidasterreno: string;
    medidasconstruccion: string;
    fechainicio: string;
    fechafin: string;
    idtipomodulo: number;
    presupuesto: number;
    idrazonsocialcliente: number;
    idrazonsocialcontratista: number;
    idrazonsocialconstructor: number;
    idrazonsocialasociado: number;
    posiciongps: string;
    idestatusmodulo: number;
    observaciones: string;
    desctipomodulo?: string;
    direccioncliente?: string;
    razonsocialcontratista?: string;
    razonsocialconstructor?: string;
    razonsocialasociado?: string;
    claveestatusmodulo?: string;
    tipomodulo?: string;
}